Bullet Updates
*Moved the pointInPoly functions to the bullet class. Also merged the
two different ones, because they were identical except for the
prototyping.

The bullet.hit(List<Point>/List<PointF>) function now will autocast to
List<PointF> if it wasn't already and run it, and it'll check for a
collision anywhere along its length. This way, it doesn't only detect
the tip, but anywhere along it. Much better collisioning!