Multi-Polygonal Enemies (and Player)
Enemies have been redesigned, rather enormously.
*All enemies now work off a List<List<PointF> > for the various polygons
that create them.

This allows enemies to have multiple, separate parts, and to have
various different colors.

+Enemies now have a List<Color> that must be filled, one entry per
polygon
+Enemies now have a List<List<PointF> > origVerts, which is the initial
shape of the enemy. The vert list will be generated on the fly at render
time.
*The enemy render function has been made non-abstract. You no longer
have to copy the code to each enemy, since it *should* be the same for
them all anyways. It's now contained in enemycontainer.cs, and calls AI
first, followed by filling out and rendering the verts list with the
colors from the colors array.

Yay, more complex enemies possible with less code to you!

Also, a new variable is passed to enemies upon being hit by a bullet -
"num," which refers to what number polygon in the array was hit. This
makes it possible to have an enemy have a wing blown off, or a shield
that absorbs hits until it fails, or separate components and each one
individually takes damage, just whatever. Enjoy!