Player Hit Function Added
Oops, meant to put this in the last update. Realized after already
commited.
+hit() function to the Player class.
*drawingpanel.cs no longer directly subtracts from the player's hp when
a bullet hits, it passes it along to the player the same way stuff is
passed to enemies when they're hit, with similar consequences and
possibilities.

This means if we were to add, for instance, a small circle powerup that
circles the player ship, if it were hit instead of the player, we could
simply have it absorb the shot and not dock the player hp. Or whatever.